Refactored readMsgHeader #14
